--- be2-scala/src/test/scala/util/examples/dataTraits/ExampleMessagesTrait.scala
@@ -1,62 +1,0 @@
-package util.examples
-
-import ch.epfl.pop.model.objects.PublicKey
-import ch.epfl.pop.model.objects.Base64Data
-import ch.epfl.pop.model.objects.Channel
-import java.nio.file.Path
-import ch.epfl.pop.model.network.MethodType.MethodType
-import java.nio.file.Files
-import ch.epfl.pop.model.network.JsonRpcRequest
-import ch.epfl.pop.model.network.method.message.data.ObjectType._
-import ch.epfl.pop.model.network.method.message.data.ActionType._
-import ch.epfl.pop.model.network.method.message.data
-
-/**
-  * Trait to be extended by other Feature traits
-  */
-trait ExampleMessagesTrait {
-  import ExampleMessages._
-
-  //Default Channel
-  val CHANNEL = DEFAULT_CHANNEL
-
-  //Object and Action type of the message
-  val obj: ObjectType
-  val action: ActionType
-
-  val METHOD_TYPE: MethodType
-
-  /**
-    *
-    * @param fileName file name of the .json file to get the payload from
-    * @param messageBuilder mid level Message to use as base for HighLevelMessage
-    * @return a parsed and decoded JsonRpcRequest
-    */
-  def getJsonRPCRequestFromFile(fileName: String)(messageBuilder: HighLevelMessageGenerator.HLMessageBuilder = DEFAULT_HL_MESSAGE_BUILDER): JsonRpcRequest = {
-    val payload = getPayloadFromFile(fileName)
-    buildJsonRpcRequest(payload)(messageBuilder)
-  }
-  /**
-    * Builds json request given a data payload and predefined high level message builder
-    * @param payload the payload to parse
-    * @param highLevelMessage high level message builder for the payload
-    * @return a parsed and decoded JsonRpcRequest
-    */
-  def buildJsonRpcRequest(payload: String)(highLevelMessage: HighLevelMessageGenerator.HLMessageBuilder): JsonRpcRequest = {
-    highLevelMessage.withMethodType(METHOD_TYPE).withChannel(CHANNEL).withPayload(payload).generateJsonRpcRequestWith(obj)(action)
-  }
-}
-
-object ExampleMessages {
-  val DEFAULT_BASE_PATH = "../protocol/examples/messageData"
-  val DEFAULT_SENDER = PublicKey(Base64Data("J9fBzJV70Jk5c-i3277Uq4CmeL4t53WDfUghaK0HpeM="))
-  val DEFAULT_CHANNEL = Channel.ROOT_CHANNEL
-
-  //Default builder used mid level Message
-  final val DEFAULT_MESSAGE_BUILDER    = new HighLevelMessageGenerator.MessageBuilder().withSender(DEFAULT_SENDER)
-  final val DEFAULT_HL_MESSAGE_BUILDER = new HighLevelMessageGenerator.HLMessageBuilder(DEFAULT_MESSAGE_BUILDER).withId(1).withChannel(DEFAULT_CHANNEL)
-
-  //Reads and returns payload from .json file
-  final def getPayloadFromFile(fileName: String): String = Files.readString(Path.of(DEFAULT_BASE_PATH,fileName))
-
-}
